requirements.txt
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
chromedriver-binary-auto==0.3.1
exceptiongroup==1.3.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
outcome==1.3.0.post0
packaging==25.0
pluggy==1.6.0
PySocks==1.7.1
pytest==8.3.3
python-dotenv==1.1.1
requests==2.32.5
selenium==4.24.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.2.1
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==1.26.20
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
